+++ ddth-dao-core/src/main/java/com/github/ddth/dao/BaseBo.java
@@ -2,14 +2,16 @@ package com.github.ddth.dao;
 
 import java.util.HashMap;
 import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.locks.Lock;
 import java.util.concurrent.locks.ReentrantLock;
 
 import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.apache.commons.lang3.builder.HashCodeBuilder;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonProperty;
+import com.github.ddth.commons.serialization.DeserializationException;
+import com.github.ddth.commons.serialization.ISerializationSupport;
+import com.github.ddth.commons.serialization.SerializationException;
 import com.github.ddth.commons.utils.DPathUtils;
 import com.github.ddth.commons.utils.SerializationUtils;
 
@@ -19,11 +21,15 @@ import com.github.ddth.commons.utils.SerializationUtils;
  * @author Thanh Ba Nguyen <bnguyen2k@gmail.com>
  * @since 0.1.0
  */
-public class BaseBo implements Cloneable {
-    @JsonProperty
-    private Map<String, Object> attributes = new HashMap<String, Object>();
+public class BaseBo implements Cloneable, ISerializationSupport {
 
-    @JsonIgnore
+    protected Map<String, Object> initAttributes(Map<String, Object> initData) {
+        return initData != null ? new ConcurrentHashMap<String, Object>(initData)
+                : new ConcurrentHashMap<String, Object>();
+    }
+
+    private Map<String, Object> attributes = initAttributes(null);
+
     private boolean dirty = false;
 
     /**
@@ -32,14 +38,17 @@ public class BaseBo implements Cloneable {
      * @since 0.5.0.5
      */
     public BaseBo clone() {
+        lock.lock();
         try {
             BaseBo obj = (BaseBo) super.clone();
-            obj.attributes = attributes != null ? new HashMap<String, Object>(attributes)
-                    : new HashMap<String, Object>();
+            obj.attributes = initAttributes(attributes);
+            obj.dirty = dirty;
             obj.lock = new ReentrantLock();
             return obj;
         } catch (CloneNotSupportedException e) {
             throw new RuntimeException(e);
+        } finally {
+            lock.unlock();
         }
     }
 
@@ -48,7 +57,6 @@ public class BaseBo implements Cloneable {
      * 
      * @return
      */
-    @JsonIgnore
     public boolean isDirty() {
         return dirty;
     }
@@ -105,9 +113,14 @@ public class BaseBo implements Cloneable {
      * @see DPathUtils
      */
     protected BaseBo setAttribute(String dPath, Object value) {
-        DPathUtils.setValue(attributes, dPath, value);
-        markDirty();
-        return this;
+        lock.lock();
+        try {
+            DPathUtils.setValue(attributes, dPath, value);
+            markDirty();
+            return this;
+        } finally {
+            lock.unlock();
+        }
     }
 
     /**
@@ -119,8 +132,10 @@ public class BaseBo implements Cloneable {
         // EMPTY
     }
 
-    @JsonIgnore
     private Lock lock = new ReentrantLock();
+
+    private final static String SER_FIELD_ATTRS = "_attrs_";
+    private final static String SER_FIELD_DIRTY = "_dirty_";
 
     /**
      * Populates the BO with data from a Java map.
@@ -128,18 +143,21 @@ public class BaseBo implements Cloneable {
      * @param data
      * @return
      */
+    @SuppressWarnings("unchecked")
     public BaseBo fromMap(Map<String, Object> data) {
-        lock.lock();
-        try {
-            attributes = new HashMap<String, Object>();
-            if (data != null) {
-                attributes.putAll(data);
+        if (data != null) {
+            lock.lock();
+            try {
+                Boolean dirty = DPathUtils.getValue(data, SER_FIELD_DIRTY, Boolean.class);
+                Map<String, Object> attrs = DPathUtils.getValue(data, SER_FIELD_ATTRS, Map.class);
+                this.attributes = initAttributes(attrs);
+                this.dirty = dirty != null ? dirty.booleanValue() : true;
+                triggerPopulate();
+            } finally {
+                lock.unlock();
             }
-            triggerPopulate();
-        } finally {
-            lock.unlock();
         }
-        return markClean();
+        return this;
     }
 
     /**
@@ -148,16 +166,16 @@ public class BaseBo implements Cloneable {
      * @return
      */
     public Map<String, Object> toMap() {
-        Map<String, Object> result = new HashMap<String, Object>();
         lock.lock();
         try {
-            if (attributes != null) {
-                result.putAll(attributes);
-            }
+            Map<String, Object> data = new HashMap<>();
+            data.put(SER_FIELD_DIRTY, dirty);
+            data.put(SER_FIELD_ATTRS,
+                    attributes != null ? new HashMap<>(attributes) : new HashMap<>());
+            return data;
         } finally {
             lock.unlock();
         }
-        return result;
     }
 
     /**
@@ -169,21 +187,14 @@ public class BaseBo implements Cloneable {
      */
     @SuppressWarnings("unchecked")
     public BaseBo fromJson(String jsonString) {
-        Map<String, Object> attrs = null;
+        Map<String, Object> dataMap = null;
         try {
-            attrs = jsonString != null ? SerializationUtils.fromJsonString(jsonString, Map.class)
+            dataMap = jsonString != null ? SerializationUtils.fromJsonString(jsonString, Map.class)
                     : null;
         } catch (Exception e) {
-            attrs = null;
+            dataMap = null;
         }
-        lock.lock();
-        try {
-            this.attributes = attrs != null ? attrs : new HashMap<String, Object>();
-            triggerPopulate();
-        } finally {
-            lock.unlock();
-        }
-        return markClean();
+        return fromMap(dataMap);
     }
 
     /**
@@ -192,12 +203,8 @@ public class BaseBo implements Cloneable {
      * @return
      */
     public String toJson() {
-        lock.lock();
-        try {
-            return SerializationUtils.toJsonString(attributes);
-        } finally {
-            lock.unlock();
-        }
+        Map<String, Object> data = toMap();
+        return SerializationUtils.toJsonString(data);
     }
 
     /**
@@ -210,20 +217,13 @@ public class BaseBo implements Cloneable {
      */
     @SuppressWarnings("unchecked")
     public BaseBo fromByteArray(byte[] data) {
-        Map<String, Object> attrs = null;
+        Map<String, Object> dataMap = null;
         try {
-            attrs = data != null ? SerializationUtils.fromByteArrayKryo(data, HashMap.class) : null;
+            dataMap = data != null ? SerializationUtils.fromByteArray(data, HashMap.class) : null;
         } catch (Exception e) {
-            attrs = null;
+            dataMap = null;
         }
-        lock.lock();
-        try {
-            this.attributes = attrs != null ? attrs : new HashMap<String, Object>();
-            triggerPopulate();
-        } finally {
-            lock.unlock();
-        }
-        return markClean();
+        return fromMap(dataMap);
     }
 
     /**
@@ -233,12 +233,8 @@ public class BaseBo implements Cloneable {
      * @since 0.5.0
      */
     public byte[] toByteArray() {
-        lock.lock();
-        try {
-            return SerializationUtils.toByteArrayKryo(attributes);
-        } finally {
-            lock.unlock();
-        }
+        Map<String, Object> dataMap = toMap();
+        return SerializationUtils.toByteArray(dataMap);
     }
 
     /**
@@ -246,23 +242,26 @@ public class BaseBo implements Cloneable {
      */
     @Override
     public boolean equals(Object obj) {
-        if (!(obj instanceof BaseBo)) {
-            return false;
+        if (obj == this) {
+            return true;
         }
-        BaseBo other = (BaseBo) obj;
-        EqualsBuilder eb = new EqualsBuilder();
-        lock.lock();
-        try {
-            other.lock.lock();
+        if (obj instanceof BaseBo) {
+            BaseBo other = (BaseBo) obj;
+            EqualsBuilder eb = new EqualsBuilder();
+            lock.lock();
             try {
-                eb.append(attributes, other.attributes);
+                other.lock.lock();
+                try {
+                    eb.append(attributes, other.attributes);
+                    return eb.isEquals();
+                } finally {
+                    other.lock.unlock();
+                }
             } finally {
-                other.lock.unlock();
+                lock.unlock();
             }
-        } finally {
-            lock.unlock();
         }
-        return eb.isEquals();
+        return false;
     }
 
     /**
@@ -287,16 +286,23 @@ public class BaseBo implements Cloneable {
         return toJson();
     }
 
-    public static void main(String[] args) {
-        BaseBo obj = new BaseBo();
-        obj.setAttribute("attr1", "value1");
-        System.out.println(obj);
+    /**
+     * {@inheritDoc}
+     * 
+     * @since 0.6.0
+     */
+    @Override
+    public byte[] toBytes() throws SerializationException {
+        return toByteArray();
+    }
 
-        BaseBo obj2 = obj.clone();
-        System.out.println(obj2);
-
-        obj.setAttribute("attr2", "value2");
-        System.out.println(obj);
-        System.out.println(obj2);
+    /**
+     * {@inheritDoc}
+     * 
+     * @since 0.6.0
+     */
+    @Override
+    public ISerializationSupport fromBytes(byte[] data) throws DeserializationException {
+        return fromByteArray(data);
     }
 }